% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDNRA-package.R
\docType{package}
\name{HDNRA-package}
\alias{HDNRA}
\alias{HDNRA-package}
\title{HDNRA: High-Dimensional Location Testing with Normal-Reference Approaches}
\description{
Provides inverse-free high-dimensional location tests for two-sample and general linear hypothesis testing (GLHT) problems under equal or unequal covariance structures. The package implements classical normal-approximation procedures, scale-invariant procedures, normal-reference procedures based on covariance-matched Gaussian companions, and F-type normal-reference calibrations for heteroscedastic Behrens-Fisher and GLHT settings. Implemented two-sample normal-approximation and scale-invariant procedures include Bai and Saranadasa (1996) \url{https://www.jstor.org/stable/24306018}, Chen and Qin (2010) \doi{10.1214/09-aos716}, Srivastava and Du (2008) \doi{10.1016/j.jmva.2006.11.002}, and Srivastava et al. (2013) \doi{10.1016/j.jmva.2012.08.014}. Implemented two-sample normal-reference procedures include Zhang, Guo, Zhou and Cheng (2020) \doi{10.1080/01621459.2019.1604366}, Zhang, Zhou, Guo and Zhu (2021) \doi{10.1016/j.jspi.2020.11.008}, Zhang, Zhu and Zhang (2020) \doi{10.1016/j.ecosta.2019.12.002}, Zhang, Zhu and Zhang (2023) \doi{10.1080/02664763.2020.1834516}, Zhang and Zhu (2022) \doi{10.1080/10485252.2021.2015768}, Zhang and Zhu (2022) \doi{10.1007/s42519-021-00232-w}, and Zhu, Wang and Zhang (2023) \doi{10.1007/s00180-023-01433-6}. Implemented GLHT normal-approximation procedures include Fujikoshi et al. (2004) \doi{10.14490/jjss.34.19}, Srivastava and Fujikoshi (2006) \doi{10.1016/j.jmva.2005.08.010}, Yamada and Srivastava (2012) \doi{10.1080/03610926.2011.581786}, Schott (2007) \doi{10.1016/j.jmva.2006.11.007}, and Zhou, Guo and Zhang (2017) \doi{10.1016/j.jspi.2017.03.005}. Implemented GLHT normal-reference procedures include Zhang, Guo and Zhou (2017) \doi{10.1016/j.jmva.2017.01.002}, Zhang, Zhou and Guo (2022) \doi{10.1016/j.jmva.2021.104816}, Zhu, Zhang and Zhang (2022) \doi{10.5705/ss.202020.0362}, Zhu and Zhang (2022) \doi{10.1007/s00180-021-01110-6}, Zhang and Zhu (2022) \doi{10.1016/j.csda.2021.107385}, and Cao et al. (2024) \doi{10.1007/s00362-024-01530-8}. The package also includes the random-integration normal-approximation GLHT procedure of Li et al. (2025) \doi{10.1007/s00362-024-01624-3}. A package-level overview is given in Wang, Zhu and Zhang (2026) \doi{10.1016/j.csda.2025.108269}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nie23wp8738/HDNRA}
  \item \url{https://nie23wp8738.github.io/HDNRA/}
  \item Report bugs at \url{https://github.com/nie23wp8738/HDNRA/issues}
}

}
\author{
\strong{Maintainer}: Pengfei Wang \email{nie23.wp8738@e.ntu.edu.sg}

Authors:
\itemize{
  \item Shuqi Luo \email{nie23.ls4909@e.ntu.edu.sg}
  \item Tianming Zhu \email{tianming.zhu@nie.edu.sg}
  \item Bu Zhou \email{bu.zhou@u.nus.edu}
}

}
\keyword{internal}
