% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.NRtest.R
\name{print.NRtest}
\alias{print.NRtest}
\title{Print Method for S3 Class "NRtest"}
\usage{
## S3 method for class \pkg{NRtest}
\method{print}{NRtest}(x, ...)
}
\arguments{
\item{x}{an \pkg{NRtest} object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints the details of the \pkg{NRtest} object in a user-friendly manner. This method
provides a clear and concise presentation of the test results contained within
the \pkg{NRtest} object, including all relevant statistical metrics and test details.
}
\details{
The \code{print.NRtest} function formats and presents the contents of the \pkg{NRtest}
object, which includes statistical test results and related parameters. This
function is designed to provide a user-friendly display of the object's
contents, making it easier to understand the results of the analysis.
}
\section{Value}{

Invisibly returns the input \code{x}.
}

\seealso{
\code{\link{NRtest.object}}
}
\author{
Pengfei Wang \email{nie23.wp8738@e.ntu.edu.sg}
}
\concept{object}
