% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WZ2026.GLHTBF.2cNRT.R
\name{WZ2026.GLHTBF.2cNRT}
\alias{WZ2026.GLHTBF.2cNRT}
\title{F-approximation-based F-type test for GLHT problem under heteroscedasticity}
\usage{
WZ2026.GLHTBF.2cNRT(Y, G, n, p)
}
\arguments{
\item{Y}{A list of \eqn{k} data matrices. The \eqn{i}th element represents the data matrix
(\eqn{n_i \times p}) from the \eqn{i}th population with each row representing a \eqn{p}-dimensional observation.}

\item{G}{A known full-rank coefficient matrix (\eqn{q \times k}) with \eqn{\operatorname{rank}(\boldsymbol{G}) < k}.}

\item{n}{A vector of \eqn{k} sample sizes. The \eqn{i}th element represents the sample size of group \eqn{i}, \eqn{n_i}.}

\item{p}{The dimension of data.}
}
\value{
A list of class \code{"NRtest"} containing the results of the hypothesis test.
}
\description{
An F-type normal reference test for the high-dimensional general linear hypothesis testing (GLHT)
problem under heteroscedasticity. The null distribution is approximated by an F distribution
using Welch--Satterthwaite (W--S) chi-square approximations.
}
\details{
The test statistic is of F-type form
\deqn{F_{n,p} = \frac{\|\boldsymbol{C\hat\mu}\|^2}{\operatorname{tr}(\widehat{\Omega}_n)}.}
The degrees of freedom are estimated by matching the first two cumulants via W--S approximation,
resulting in an \eqn{F_{\hat d_1, \hat d_2}} reference distribution.
}
\examples{
library("HDNRA")
data("corneal")
dim(corneal)
group1 <- as.matrix(corneal[1:43, ]) ## normal group
group2 <- as.matrix(corneal[44:57, ]) ## unilateral suspect group
group3 <- as.matrix(corneal[58:78, ]) ## suspect map group
group4 <- as.matrix(corneal[79:150, ]) ## clinical keratoconus group
p <- dim(corneal)[2]
Y <- list()
k <- 4
Y[[1]] <- group1
Y[[2]] <- group2
Y[[3]] <- group3
Y[[4]] <- group4
n <- c(nrow(Y[[1]]),nrow(Y[[2]]),nrow(Y[[3]]),nrow(Y[[4]]))
G <- cbind(diag(k-1),rep(-1,k-1))
WZ2026.GLHTBF.2cNRT(Y,G,n,p)


}
\references{
Wang, P. and Zhu, T. (preprint). An F-type Test for Heteroscedastic General Linear Hypothesis Testing
Problem for High Dimensional Data: a Normal Reference Approach.
}
\concept{glht}
